+++ team4_FinalProj/src/main/java/com/team4/finalproj/order/OrderCompleteController.java
@@ -1,12 +1,12 @@
-package com.team4.finalproj;
+package com.team4.finalproj.order;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-public class OrderProcessController {
-	@RequestMapping("orderprocess")
+public class OrderCompleteController {
+	@RequestMapping("ordercomplete")
 	public String orderProcess(){
-		return "orderprocess";
+		return "ordercomplete";
 	}
 }
