+++ team4_FinalProj/src/main/java/com/team4/finalproj/order/OrderProcessController.java
@@ -1,12 +1,12 @@
-package com.team4.finalproj;
+package com.team4.finalproj.order;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-public class OrderCompleteController {
-	@RequestMapping("ordercomplete")
+public class OrderProcessController {
+	@RequestMapping("orderprocess")
 	public String orderProcess(){
-		return "ordercomplete";
+		return "orderprocess";
 	}
 }
