--- team4_FinalProj/src/main/java/com/team4/model/EventDto.java
@@ -1,5 +1,0 @@
-package com.team4.model;
-
-public class EventDto {
-
-}
